+++ records/urls.py
@@ -1,7 +1,6 @@
 from django.urls import path
 from . import views
 
-
 urlpatterns = [path('', views.records, name='records'),
                path('record/<str:pk>/', views.record, name='record'),
                path('create-record/', views.create_record, name='create-record'),
@@ -10,4 +9,6 @@ urlpatterns = [path('', views.records, name='records'),
                path('reports/', views.get_reports, name='reports'),
                path('export-to-csv/', views.export_to_csv, name='export-to-csv'),
                path('export/', views.export_to_csv_view, name='export'),
+               path('export-to-pdf/', views.export_to_pdf, name='export-pdf'),
+               path('export-pdf-view/', views.export_to_pdf_view, name='export-pdf-view'),
                ]