+++ waste/settings.py
@@ -5,9 +5,9 @@ from datetime import timedelta
 
 BASE_DIR = Path(__file__).resolve().parent.parent
 
-DEBUG = False
+DEBUG = True
 
-ALLOWED_HOSTS = ['127.0.0.1']
+ALLOWED_HOSTS = ["localhost", "127.0.0.1"]
 
 # CSRF_TRUSTED_ORIGINS = ['']